Move the component into the fields folder
